Calculation Refactor: - changed all double to BigDecimal for a more precise result, and updated all dependencies. - removed currentPrice, averagePrice, yearLow, yearHigh, change, and percentChange from Ticker. - changed resetting of EditText from "onDestroyView" to "onStop" for all calculator tabs. -- consider saving instead of resetting. - added permission for INTERNET and ACCESS_NETWORK_STATE - added empty HttpClient interface - added packagingOptions to fix Android duplicate exception - added retrofit and mockito imports - added retrofit to proguard-rules.pro

+++ app/src/main/java/com/aaron/pseplanner/adapter/TickerListAdapter.java
@@ -86,14 +86,14 @@ public class TickerListAdapter extends ArrayAdapter<Ticker>
         {
             layout.setOnTouchListener(listener);
             stock.setText(ticker.getSymbol());
-            price.setText(service.formatStockPrice(ticker.getCurrentPrice()));
-            change.setText(service.formatStockPrice(ticker.getChange()));
-            String percentChangeText = service.formatStockPrice(ticker.getPercentChange()) + "%";
+            price.setText(service.formatStockPrice(ticker.getCurrentPrice().doubleValue()));
+            change.setText(service.formatStockPrice(ticker.getChange().doubleValue()));
+            String percentChangeText = service.formatStockPrice(ticker.getPercentChange().doubleValue()) + "%";
             percentChange.setText(percentChangeText);
 
-            service.formatTextColor(ticker.getChange(), price);
-            service.formatTextColor(ticker.getChange(), change);
-            service.formatTextColor(ticker.getChange(), percentChange);
+            service.formatTextColor(ticker.getChange().doubleValue(), price);
+            service.formatTextColor(ticker.getChange().doubleValue(), change);
+            service.formatTextColor(ticker.getChange().doubleValue(), percentChange);
         }
     }
 }
